+++ lib/chartkick.rb
@@ -6,5 +6,7 @@ require "chartkick/sinatra" if defined?(Sinatra)
 module Chartkick
   class << self
     attr_accessor :content_for
+    attr_accessor :options
   end
+  self.options = {}
 end